end part 17.

+++ Matrix4x4.h
@@ -97,10 +97,10 @@ public:
 			}
 			v.cross(vec[0], vec[1], vec[2]);
 
-			out.m_mat[0][i] = pow(-1.0f, i) * v.m_x / det;
-			out.m_mat[1][i] = pow(-1.0f, i) * v.m_y / det;
-			out.m_mat[2][i] = pow(-1.0f, i) * v.m_z / det;
-			out.m_mat[3][i] = pow(-1.0f, i) * v.m_w / det;
+			out.m_mat[0][i] = static_cast<float>(pow(-1.0f, i) * v.m_x / det);
+			out.m_mat[1][i] = static_cast<float>(pow(-1.0f, i) * v.m_y / det);
+			out.m_mat[2][i] = static_cast<float>(pow(-1.0f, i) * v.m_z / det);
+			out.m_mat[3][i] = static_cast<float>(pow(-1.0f, i) * v.m_w / det);
 		}
 
 		this->setMatrix(out);
